Remove DataFrameTable.__init__
This allows for the full set of DataTable initialisation arguments to be
passed through via DataFrameTable. See

    https://textual.textualize.io/widgets/data_table/#textual.widgets.DataTable

for the parameters that people may wish to be using. This means that
DataFrameTable is now a drop-in replacement for DataTable.

If DataFrameTable does need to grow its own __init__ it would be a good idea
to carry all of those arguments through.